Move net.sf.jabref.TestUtils to package net.sf.jabref.testutils and rename to GuiTestUtils
